requirements.txt
Flask>=3.0,<4.0
Flask-SQLAlchemy>=3.1,<4.0
Flask-Migrate>=4.0,<5.0
Flask-Login>=0.6,<1.0
Werkzeug>=3.0,<4.0
SQLAlchemy>=2.0,<3.0
python-dateutil>=2.9,<3.0
stripe>=10.8,<11.0
python-dotenv>=1.0,<2.0
gunicorn>=23.0,<24.0
Jinja2>=3.1.6,<4.0
itsdangerous>=2.2,<3.0
click>=8.2,<9.0
blinker>=1.9,<2.0
alembic>=1.16,<2.0
MarkupSafe>=3.0,<4.0
typing-extensions>=4.15,<5.0
requests>=2.32,<3.0
Mako>=1.3,<2.0
packaging>=25.0,<26.0
charset-normalizer>=3.4,<4.0
idna>=3.10,<4.0
urllib3>=2.5,<3.0
certifi>=2025.8,<2026.0
six>=1.17,<2.0
